feat: move model.py src/api/schemas
